Refactor  timestamp microservice
