--- backend-dev/projects/file-metadata-microservice/package.json
@@ -1,35 +1,0 @@
-{
-  "name": "file-metadata-microservice",
-  "version": "1.0.0",
-  "description": "Extract's metadata from uploaded file. Project from FreeCodeCamp's Backend Development and APIs section.",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ChristianLapinig/fcc.git"
-  },
-  "keywords": [
-    "fcc",
-    "freecodecamp",
-    "nodejs",
-    "express",
-    "file",
-    "metadata"
-  ],
-  "author": "Christian Lapinig",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/ChristianLapinig/fcc/issues"
-  },
-  "homepage": "https://github.com/ChristianLapinig/fcc#readme",
-  "dependencies": {
-    "express": "^4.18.2",
-    "multer": "^1.4.5-lts.1"
-  },
-  "devDependencies": {
-    "nodemon": "^3.0.2"
-  }
-}
