+++ backend-dev/projects/router.js
@@ -1,0 +1,6 @@
+const express = require("express");
+const router = express.Router();
+
+router.use("/", require("./timestamp-microservice/timestamp"));
+
+module.exports = router;
