--- backend-dev/projects/timestamp-microservice/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "timestamp-microservice",
-  "version": "1.0.0",
-  "description": "FreeCodeCamp Timestamp Microservice project for backend developement\"",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ChristianLapinig/fcc.git"
-  },
-  "keywords": [
-    "fcc",
-    "microservice",
-    "timestamp"
-  ],
-  "author": "Christian Lapinig",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/ChristianLapinig/fcc/issues"
-  },
-  "homepage": "https://github.com/ChristianLapinig/fcc#readme",
-  "dependencies": {
-    "express": "^4.18.2"
-  },
-  "devDependencies": {
-    "nodemon": "^3.0.2"
-  }
-}
