--- backend-dev/projects/url-shortener/package.json
@@ -1,37 +1,0 @@
-{
-  "name": "url-shortener",
-  "version": "1.0.0",
-  "description": "URL shortener project from FreeCodeCamp's backend developer curriculum",
-  "main": "index.js",
-  "scripts": {
-    "start": "node index.js",
-    "dev": "nodemon index.js"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/ChristianLapinig/fcc.git"
-  },
-  "keywords": [
-    "fcc",
-    "url",
-    "shortener",
-    "nodejs",
-    "express"
-  ],
-  "author": "Christian Lapinig",
-  "license": "ISC",
-  "bugs": {
-    "url": "https://github.com/ChristianLapinig/fcc/issues"
-  },
-  "homepage": "https://github.com/ChristianLapinig/fcc#readme",
-  "dependencies": {
-    "express": "^4.18.2",
-    "mongoose": "^8.0.3",
-    "nanoid": "^5.0.4",
-    "short-unique-id": "^5.0.3"
-  },
-  "devDependencies": {
-    "dotenv": "^16.3.1",
-    "nodemon": "^3.0.2"
-  }
-}
